Agentic Ai/Schedule_Gen-v1/pyproject.toml
[project]
name = "schedule-gen-v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google-genai>=1.34.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-pinecone>=0.2.12",
    "pinecone>=7.3.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "werkzeug>=3.1.3",
]
